Update README user guidance
ADDED: Explanation about the the data file.
BECAUSE: The typos, endonyms, synonyms etc. in the file are not a definitive list and users can amend. 

CHANGED: Links to register team support pages, rather than generic email address.
BECAUSE: To provide better user support journey. 

CHANGED: Title from 'open register location picker' to just 'location picker'
BECAUSE: We don't refer to 'open registers' anymore and it's clearer for the user.